requirements.txt
python-dotenv==1.0.1
streamlit==1.37.1
langchain==0.2.14
langchain-openai==0.1.22
langchain-anthropic==0.1.23
pandas==2.2.2